backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.34.3
vaderSentiment==3.3.2
python-dotenv==1.1.1
Flask
flask-cors
psycopg2-binary